pyproject.toml
[project]
name = "asmart-test-task"
version = "0.1.0"
description = ""
authors = [
    {name = "xaoriGit",email = "xaorinit@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.8,<6.0.0)",
    "django-ninja (>=1.4.5,<2.0.0)",
    "django-ninja-jwt (>=5.4.0,<6.0.0)",
    "faker (>=37.12.0,<38.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "uvicorn (>=0.38.0,<0.39.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
